feat: add password toggle in change password page in account page
