+++ src/main/webapp/css/account.css
@@ -88,6 +88,7 @@
 .account-form .form-group {
     flex: 1;
     margin-bottom: 20px;
+    position: relative;
 }
 
 .account-form label {
@@ -140,4 +141,25 @@
 }
 .btn-primary:hover {
     background-color: blue;
+}
+
+.account-form .form-group i {
+    position: absolute;
+    top: 42px;
+    font-size: 18px;
+    z-index: 10;
+}
+
+.account-form .form-group i.bx-lock-alt {
+    left: 15px;
+}
+
+.account-form .form-group i.toggle-password {
+    right: 15px;
+    cursor: pointer;
+    transition: color 0.3s ease;
+}
+
+.account-form .form-group i.toggle-password:hover {
+    color: var(--primary-color);
 }
