+++ src/main/webapp/js/login.js
@@ -39,20 +39,6 @@ function register() {
     adjustHeight(true);
 }
 
-window.togglePassword = function (inputId, icon) {
-    const input = document.getElementById(inputId);
-
-    if (input.type === "password") {
-        input.type = "text";
-        icon.classList.remove("fa-eye-slash")
-        icon.classList.add("fa-eye")
-    } else {
-        input.type = "password";
-        icon.classList.remove("fa-eye");
-        icon.classList.add("fa-eye-slash");
-    }
-}
-
 export function initializeLogin() {
     const loginTrigger = document.getElementById('login-trigger');
     const registerTrigger = document.getElementById('register-trigger');
